// src/interFace/interFace.ts
export interface IMissionVision {
  id: number;
  img: StaticImageData;
  title: string;
  description: string;
}